exercise_04_version2

+++ ani.ter-markosyan/repo/Lesson_3/exercise_04/exercise_04.py
@@ -1,2 +1,3 @@
 #!usr/bin/evn python3
-print(round((float((512 - 282) / ((47 * 48) + 5))), 4))
+
+print(round(((512 - 282) / ((47 * 48) + 5)), 4))
